domain 수정

+++ src/main/java/dku/capstone/foodlog/domain/Post.java
@@ -28,7 +28,7 @@ public class Post extends BaseTime{
     @OneToMany(mappedBy = "post")
     private List<PostPicture> pictureList = new ArrayList<>();
 
-    private Float rating;
+    private int rating;
 
     private String review;
